lseg_market_agent/requirements.txt
google-adk
python-dotenv
requests
google-genai
mcp
beautifulsoup4